Use size_t instead of unsigned in a few places. This removes compiler warnings for use on 64 bit platforms.
Add some ATTR_UNUSED tags to method parameters that were not being used. The
goal being to use the -Werror flag for compilation of WALi. Only 1 file is
preventing this now.


git-svn-id: file:///p/wpis/people/students/kidd/svnwali/trunk@306 dfa36f5b-b2fa-0310-9605-e5aca7e64462

+++ Source/wali/graph/RegExp.hpp
@@ -57,13 +57,13 @@ namespace wali {
 
         struct hash_reg_exp_key {
             size_t operator() (const reg_exp_key_t &k) const {
-                return ((unsigned)k.type ^ (unsigned)k.c1.get_ptr() ^ (unsigned)k.c2.get_ptr());
+                return ((size_t)k.type ^ (size_t)k.c1.get_ptr() ^ (size_t)k.c2.get_ptr());
             }
         };
 
         struct hash_sem_elem {
             size_t operator() (const sem_elem_t &se) const {
-                return (unsigned)se.get_ptr();
+                return (size_t)se.get_ptr();
             }
         };
